🐛 응답 스키마 수정
