+++ src/shared/api/entities/schemas.ts
@@ -1,7 +1,16 @@
+// DTO
+type UserBriefDTO = {
+  id: string;
+  username: string;
+  isAdmin: boolean;
+};
+
+// Params
 export type EchoParams = {
   message: string;
 };
 
+// Request
 export type PretotypeUserSubmitRequest = {
   email: string;
   isSubscribed: boolean;
@@ -26,10 +35,20 @@ export type SocialSignUpRequest = {
   authProvider: string;
 };
 
+export type SendEmailCodeRequest = {
+  snuMail: string;
+};
+
+export type EmailVerifyRequest = {
+  snuMail: string;
+  code: string;
+};
+
 export type GoogleSignInRequest = {
   googleAccessToken: string;
 };
 
+// Response
 export type PretotypeUserSubmitResponse = {
   email: string;
   isSubscribed: boolean;
@@ -37,12 +56,6 @@ export type PretotypeUserSubmitResponse = {
 };
 
 export type UserWithTokenResponse = {
-  userResponse: {
-    id: BigInteger;
-    name: string;
-    email: string;
-    phoneNumber: string;
-  };
+  user: UserBriefDTO;
   accessToken: string;
-  refreshToken: string;
 };
